OneDrive/Documents/rag-ticket-bot/requirements.txt
streamlit
pandas
faiss-cpu
python-dotenv

langchain
langchain-core
langchain-community
langchain-openai
openai
tiktoken